+++ spec/grape/exceptions/unknown_options_spec.rb
@@ -1,10 +1,9 @@
-# encoding: utf-8
 require 'spec_helper'
 
 describe Grape::Exceptions::UnknownOptions do
   describe '#message' do
     let(:error) do
-      described_class.new([:a, :b])
+      described_class.new(%i[a b])
     end
 
     it 'contains the problem in the message' do