+++ spec/grape/util/reverse_stackable_values_spec.rb
@@ -9,7 +9,7 @@ module Grape
         it 'returns all keys' do
           subject[:some_thing] = :foo_bar
           subject[:some_thing_else] = :foo_bar
-          expect(subject.keys).to eq [:some_thing, :some_thing_else].sort
+          expect(subject.keys).to eq %i[some_thing some_thing_else].sort
         end
 
         it 'returns merged keys with parent' do
@@ -19,7 +19,7 @@ module Grape
           subject[:some_thing] = :foo_bar
           subject[:some_thing_more] = :foo_bar
 
-          expect(subject.keys).to eq [:some_thing, :some_thing_else, :some_thing_more].sort
+          expect(subject.keys).to eq %i[some_thing some_thing_else some_thing_more].sort
         end
       end
 
@@ -52,7 +52,7 @@ module Grape
         it 'combines parent and actual values (actual first)' do
           parent[:some_thing] = :foo
           subject[:some_thing] = :foo_bar
-          expect(subject[:some_thing]).to eq [:foo_bar, :foo]
+          expect(subject[:some_thing]).to eq %i[foo_bar foo]
         end
 
         it 'parent values are not changed' do
@@ -71,28 +71,28 @@ module Grape
         it 'pushes further values' do
           subject[:some_thing] = :foo
           subject[:some_thing] = :bar
-          expect(subject[:some_thing]).to eq [:foo, :bar]
+          expect(subject[:some_thing]).to eq %i[foo bar]
         end
 
         it 'can handle array values' do
           subject[:some_thing] = :foo
-          subject[:some_thing] = [:bar, :more]
-          expect(subject[:some_thing]).to eq [:foo, [:bar, :more]]
+          subject[:some_thing] = %i[bar more]
+          expect(subject[:some_thing]).to eq [:foo, %i[bar more]]
 
-          parent[:some_thing_else] = [:foo, :bar]
-          subject[:some_thing_else] = [:some, :bar, :foo]
+          parent[:some_thing_else] = %i[foo bar]
+          subject[:some_thing_else] = %i[some bar foo]
 
-          expect(subject[:some_thing_else]).to eq [[:some, :bar, :foo], [:foo, :bar]]
+          expect(subject[:some_thing_else]).to eq [%i[some bar foo], %i[foo bar]]
         end
       end
 
       describe '#to_hash' do
         it 'returns a Hash representation' do
           parent[:some_thing] = :foo
-          subject[:some_thing] = [:bar, :more]
+          subject[:some_thing] = %i[bar more]
           subject[:some_thing_more] = :foo_bar
           expect(subject.to_hash).to eq(
-            some_thing: [[:bar, :more], :foo],
+            some_thing: [%i[bar more], :foo],
             some_thing_more: [:foo_bar]
           )
         end
@@ -106,12 +106,12 @@ module Grape
           grandchild = described_class.new child
 
           parent[:some_thing] = :foo
-          child[:some_thing] = [:bar, :more]
+          child[:some_thing] = %i[bar more]
           grandchild[:some_thing] = :grand_foo_bar
           grandchild[:some_thing_more] = :foo_bar
 
           expect(grandchild.clone.to_hash).to eq(
-            some_thing: [:grand_foo_bar, [:bar, :more], :foo],
+            some_thing: [:grand_foo_bar, %i[bar more], :foo],
             some_thing_more: [:foo_bar]
           )
         end